# PERF: pass through to numpy validation for iloc setitem