@jbrockmendel 2020-02-26
Removed the tuple length validation, picked up on the order of a microsecond compared to the posts from the OP.